requirements.txt
torch==2.1.0
transformers==4.40.0
sentence-transformers==2.2.2
pillow==10.0.0
numpy==1.27.0
pandas==2.1.0
scikit-learn==1.4.0
flask==3.0.0